pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdf-nmpc"
version = "1.0.0"
description = "Neural SDF-based MPC for Collision Avoidance"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Martin Jacquet" }]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10.1",
    "simpy>=1.9",
    "casadi>=3.5.0",
    "pyyaml>=5.4.1",
    "h5py>=3.9.0",
    "warp-lang==1.3.1",
    "torch==2.1.0",
    "torchvision==0.16.0",
    "torchinfo>=1.8.0",
    # "l4casadi",  # to be installed manually with --no-build-isolation
    # "acados-template",  # to be installed manually from acados
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sdf_nmpc*"]
exclude = ["scripts*"]

[tool.setuptools.package-data]
sdf_nmpc = ["config/*.yaml", "data/*.pt"]
